* moved most FAISS_ASSERT calls to C++ exceptions, and adjusted   memory allocation to avoid mem leaks
* added an IndexIVFScalarQuantizer type that offers an
  intermediate compression between IVFFlat and IVFPQ

* support removal of indices in IndexIDMap / IndexFlat combination

* various fixes in GPU code

+++ gpu/test/TestUtils.h
@@ -1,4 +1,3 @@
-
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
  * All rights reserved.
@@ -55,7 +54,7 @@ T randSelect(std::initializer_list<T> vals) {
 }
 
 /// Generates a collection of random vectors in the range [0, 1]
-std::vector<float> randVecs(int num, int dim);
+std::vector<float> randVecs(size_t num, size_t dim);
 
 /// Compare two indices via query for similarity
 void compareIndices(faiss::Index& refIndex,
